influxdb-middleware/requirements.txt
paho-mqtt==1.5.1
influxdb-client